Add GlobalCauseTracker for handling tracking phases in all worlds, even those loaded during a global phase.
