+++ src/main/java/org/spongepowered/common/scheduler/SyncScheduler.java
@@ -24,15 +24,10 @@
  */
 package org.spongepowered.common.scheduler;
 
-import net.minecraft.world.WorldServer;
 import org.spongepowered.api.event.cause.NamedCause;
-import org.spongepowered.common.event.InternalNamedCauses;
-import org.spongepowered.common.event.tracking.CauseTracker;
+import org.spongepowered.common.event.tracking.GlobalCauseTracker;
 import org.spongepowered.common.event.tracking.PhaseContext;
-import org.spongepowered.common.event.tracking.phase.general.GeneralPhase;
 import org.spongepowered.common.event.tracking.phase.plugin.PluginPhase;
-import org.spongepowered.common.interfaces.world.IMixinWorldServer;
-import org.spongepowered.common.world.WorldManager;
 
 public class SyncScheduler extends SchedulerBase {
 
@@ -73,18 +68,13 @@ public class SyncScheduler extends SchedulerBase {
 
     @Override
     protected void executeTaskRunnable(ScheduledTask task, Runnable runnable) {
-        for (WorldServer worldServer : WorldManager.getWorlds()) {
-            final CauseTracker otherCauseTracker = ((IMixinWorldServer) worldServer).getCauseTracker();
-            otherCauseTracker.switchToPhase(PluginPhase.State.SCHEDULED_TASK, PhaseContext.start()
-                    .add(NamedCause.source(task))
-                    .addCaptures()
-                    .complete()
-            );
-        }
+        GlobalCauseTracker.getInstance().switchToPhase(PluginPhase.State.SCHEDULED_TASK, PhaseContext.start()
+                .add(NamedCause.source(task))
+                .addCaptures()
+                .complete()
+        );
         runnable.run();
-        for (WorldServer worldServer : WorldManager.getWorlds()) {
-            ((IMixinWorldServer) worldServer).getCauseTracker().completePhase(PluginPhase.State.SCHEDULED_TASK);
-        }
+        GlobalCauseTracker.getInstance().completePhase(PluginPhase.State.SCHEDULED_TASK);
     }
 
 }
